Add CartesianPlane and Point.
A cartesian plane has limits (x and y axis) and is
inteded to hold the game state.

A point in the cartesian plane is stored in an
SplFileObjectStorage so we can manage what is holded
into any given point, in my mind, that "bag" holding
stuff (who own a point on the plane, what it has) for
a point should be called "Tile".