Tests (integration) end game results.
Favoring always Player 1, this transposes the relevant contents of
`playgame.php` to an integration test in order to provide refactor
capabilities.

Player 1 is always favored because he moves later, overwriting the move
from player 0 on game turns. He rarely loses, for now, I think this test
should survive the ocasional failures (they are rare) as it is our only
end-to-end test.