Create `Turn` class to control game turns.
A turn is observable as a subject, the game is an observer of it and is
updated each turn occours.

The `Turn` class is also an iterator, each iteration increases one turn
util the limit of turns is reached. Every turn notifies the observers.

The Game class is made iterable, returning the Turn object of the game
so we can do update the game state every turn.