add the concept of feedback
Feedback from Score Strategies is not boolean anymore

return a Feedback where can visit the Win|Lose|Neutrilized moves for that round

Each Score strategy return its own feedback and the Arena merge the data into
a new feedback.
The Player receive this feedback and can use this to understant the results

the Feedback Constants are
WIN: when the warrior win the position
LOSE: when it lose
NEUTRALIZED: when the warrior found a enemy on the position and kill but die too.